rename to jsc for consistency
