+++ yt_dlp/extractor/youtube/jsc/__init__.py
