+++ yt_dlp/extractor/youtube/jsc/_registry.py
